requirement.txt
boto3==1.34.84
PyYAML==6.0.1